Добавление функционала: 1. У пользователя теперь имеется Имя, Фамилия, Тег 2. Добавлены кастомные Claim для Identity 3. Добавление/удаление из друзей других пользователей 4. Валидация модели при регистрации/авторизации 5. Проработан раздел "комментарии" 6. Просмотр профилей других пользователей

+++ Web.Application/CQRS/Commands/Comments/DeleteComment/DeleteCommentCommandHandler.cs
@@ -11,14 +11,18 @@ public class DeleteCommentCommandHandler(IDbContext dbContext) : IRequestHandler
     {
         var comment = await _dbContext.Comments
             .Include(x => x.User)
-            .Where(x => x.Id == request.CommentId && x.User.Id == request.UserId)
+            .Include(x => x.Post)
+            .Where(x => x.Id == request.CommentId)
             .FirstOrDefaultAsync(cancellationToken);
 
         if (comment is null)
             return false;
 
-        _dbContext.Comments.Remove(comment);
-        await _dbContext.SaveChangesAsync(cancellationToken);
+        if (comment.Post.Id == request.PostId || comment.User.UserTag == request.UserTag)
+        {
+            _dbContext.Comments.Remove(comment);
+            await _dbContext.SaveChangesAsync(cancellationToken);
+        }
 
         return true;
     }
